Pull `libcamera` heplers into its own module (#22)
Gather the libcamera centric conversion functions into one place.
